refactor: rename tpl.go to main.go
